+++ src/main/java/bowling/domain/state/Open.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.state;
 
 public class Open implements State {
