+++ src/main/java/bowling/domain/state/Pitching.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.state;
 
 public class Pitching implements State {
