+++ src/main/java/bowling/domain/state/Strike.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.state;
 
 public class Strike implements State {
