+++ src/test/java/bowling/domain/state/PinsTest.java
@@ -1,9 +1,8 @@
-package bowling.domain;
+package bowling.domain.state;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import bowling.domain.Pins;
 import org.junit.jupiter.api.Test;
 
 class PinsTest {
