+++ src/test/java/bowling/domain/state/StateTest.java
@@ -1,12 +1,7 @@
-package bowling.domain;
+package bowling.domain.state;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import bowling.domain.Open;
-import bowling.domain.Pitching;
-import bowling.domain.Spare;
-import bowling.domain.State;
-import bowling.domain.Strike;
 import org.junit.jupiter.api.Test;
 
 class StateTest {
